Move matching data exporter into namespace
