+++ spec/services/ucas_matching/matching_data_export_spec.rb
@@ -1,11 +1,11 @@
 require 'rails_helper'
 
-RSpec.describe ApplicationsExportForUCAS do
+RSpec.describe UCASMatching::MatchingDataExport do
   let(:unsubmitted_form) { create(:application_form) }
   let(:submitted_form) { create(:completed_application_form, application_choices_count: 3) }
 
   describe '#relevant_applications' do
-    let(:result) { ApplicationsExportForUCAS.new.send(:relevant_applications) }
+    let(:result) { described_class.new.send(:relevant_applications) }
 
     it 'includes ApplicationForms which are submitted' do
       expect(result).to include(submitted_form)
@@ -22,7 +22,7 @@ RSpec.describe ApplicationsExportForUCAS do
   end
 
   describe '#applications' do
-    let(:result) { ApplicationsExportForUCAS.new.applications }
+    let(:result) { described_class.new.applications }
 
     it 'returns an array' do
       expect(result).to be_a(Array)
